+++ src/SJP.Avro.Tools.CodeGen.Tests/SyntaxUtilitiesTests.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using NUnit.Framework;
 
 namespace SJP.Avro.Tools.CodeGen.Tests;
@@ -24,5 +24,44 @@ internal static class SyntaxUtilitiesTests
     public static void SafeNamespaceName_GivenMissingOrEmptySegment_ThrowsArgumentException(string input)
     {
         Assert.That(() => SyntaxUtilities.SafeNamespaceName(input), Throws.InstanceOf<ArgumentException>());
+    }
+
+    [TestCase(null)]
+    [TestCase("")]
+    [TestCase("  ")]
+    [TestCase("\n\n")]
+    [TestCase("*")]
+    [TestCase("*\n*\n*")]
+    public static void BuildCommentTrivia_GivenCommentWithoutText_ReturnsNoTrivia(string comment)
+    {
+        Assert.That(SyntaxUtilities.BuildCommentTrivia(comment), Is.Empty);
+    }
+
+    [Test]
+    public static void BuildCommentTrivia_GivenSingleLineComment_ReturnsSummaryContainingComment()
+    {
+        var trivia = SyntaxUtilities.BuildCommentTrivia("A test comment.").ToFullString();
+
+        Assert.That(trivia, Does.Contain("<summary>").And.Contain("A test comment.").And.Contain("</summary>"));
+    }
+
+    [Test]
+    public static void BuildCommentTrivia_GivenCommentWithXmlSyntax_EscapesXmlSyntax()
+    {
+        var trivia = SyntaxUtilities.BuildCommentTrivia("a < b & c ]]> d").ToFullString();
+
+        Assert.That(trivia, Does.Contain("a &lt; b &amp; c ]]&gt; d"));
+    }
+
+    [Test]
+    public static void BuildCommentTrivia_GivenMultipleParagraphs_ReturnsParagraphPerBlankSeparatedBlock()
+    {
+        var trivia = SyntaxUtilities.BuildCommentTrivia("first\n \nsecond").ToFullString();
+
+        Assert.Multiple(() =>
+        {
+            Assert.That(trivia, Does.Contain("<para>first</para>"));
+            Assert.That(trivia, Does.Contain("<para>second</para>"));
+        });
     }
 }